Alternative way of looking up Discourse users
This is the first step, so far only used for comments, where it is urgently needed. Will be used for posts as well later on. The primary issue is that some users have posted comments although they have not made posts, which means they are missing from the initial fetching. So we need some way of creating Discourse users on the fly as we find new users in the comments. A second issue is when doing separate import runs for different groups. We then need to check not only if a user was one of the posters in the current batch, but also if the user exists in the database due to a previous import.