Skip validations when creating posts
This is a crucial step when importing a large number of posts but I did not discover it until late in my work. I would have saved a lot of time if I had found it earlier. Thank you @SamSaffron for giving me the pointer when I finally asked on the Discourse dev forum!