Rework Discourse user creation process
This is a somewhat major update so let me explain what is going on. The previous version had a number of problems, most importantly it used names in a rather naive way to create Discourse user names. As a consequence, if two users had the same name, one of them got all their posts under the special user called Unknown. Since it is actually common in groups with a few hundred members for two of them to have the same name, e.g. John Smith, this was necessary for me to fix.

One important step is to generate better usernames, adding a number and/or incrementing it in cases of duplicates. This way, any number of users can ha the same name, they will simply get usernames JohnSmith, JohnSmith2, JohnSmith3 etc. I also use Facebook IDs for lookups instead of names and have done away with the special hash of imported users. This was obsolete anyway since I am reworking the script to support sequential imports from multiple groups, so I need to be able to lookup users imported in previous sessions. See commit 1e96166 for details and note that this commit replaces some of mechanisms implemented there.

One consequence of the above is that the script no longer does an initial batch import of users but rather creates them on the fly as needed while importing posts and comments. The functions for importing posts and comments respectively sends the from field of Facebook objects to a new method which then does lookup based in Facebook ID and then if necessary calls the previously existing method for creating a new user account.

The creation and usage of a special UnknownUser object is formalized in a function of its own, to either look it up or create it.