Import likes
This was simpler than I had thought. Just ask the API to include like_count when fetching comments and then add a step to the post and comment importing functions. I had to disable Discourse rate limiting to make this work, since imported likes are registered so quickly in sequence.