Option for slowing down API access to avoid rate limiting
This is not needed for importing small groups, but I found it necessary for medium to large size group. Since the importer has become more ambitious (likes, replies to comments) I am adding this with a default setting of 1 second delay before each API call. It might be possible to set the delay a bit lower but in any case, with this setting it is almost impossible to exceed the normal rate limits of the Graph API. Note that some calls are not delayed, e.g. paging through results.