Optional mode where fetched data are stored to the file system
This is useful for debugging imports of large scripts and also generates a JSON archive of all exported data. One use for this is to export data to one machine (e.g. development) and then transfer the files to import on another machine (e.g. production) without downloading everything an additional time.