Better exiting routine
Most importantly this ensures that Discourse site settings are always restored, even if the script crashes. It also prints a report of everything done by the script, including errors and anomalies, every time the script ends. Note that it works just fine to exit by pressing ctrl-c, the report will still be printed.