Rework the test mode
Since the importer now does so much more than the original version, the old test mode was a bit obsolete. So I made a new one. It is somewhat quick and dirty but it is actually very useful since it fetches all the data (posts, comments, likes, images etc) and if store_data_to_files is set to true, an actual run after the test run does not need to refetch the data.