Enable offline importing of data stored in JSON files
These are the final steps after preparations in previous commits. The API connection is now only initialized when first called, which should be never with a complete JSON archive. Other than that it is ensured that test mode actually creates all user objects needed for importing likes and tags and that the importer does not attempt to fetch likes for posts or comments without like counts.