Remove redundant site setting modifications
Since we are now disabling the RateLimiter as well as passing :skip_validations to the post creation method, we do not need these anymore. I am doubtful that we need any of the site settings changes except disable_emails but I need to do more testing before I remove the rest.

The problem with having redundant settings modifications is that restoring the original settings does not always work, i.e. if the script crashes before the restore method is run. Even though we are now using begin...ensure for this, it can still happen (happened to me due to a Redis related issue in a production environment).