Remove SiteSetting backup and restore routine
As far as I can tell, these are not needed. I have tested extensively but not experienced any issues, even when exceeding the limits in question. Since there is a risk of messing things up if settings are not restored (if the import script crashes), I am removing them now.