Store empty lists to disk for nonfetchable likes
This handles a special case which occurs for comments posted by users who have since shut down their Facebook accounts, which leads to the comments being visible and having like counts, but the API throws an error when trying to fetch the likes.

Perhaps this solution does not look very elegant but the only alternative I have come up with is adding a special mode for the importer, 'only import from files, never access API' mode. I actually tried this out but it became very messy.